fix: readd dropdown menu to cells and add more options
