+++ apps/webservice/src/app/[workspaceSlug]/(app)/(deploy)/(raw)/systems/[systemSlug]/(raw)/deployments/[deploymentSlug]/(raw)/releases/[releaseId]/jobs/release-table/ResourceReleaseTable.tsx
@@ -195,6 +195,7 @@ const CollapsibleTableRow: React.FC<CollapsibleTableRowProps> = ({
                   key={approval.id}
                   approval={approval}
                   deploymentVersion={deploymentVersion}
+                  environmentId={environment.id}
                 />
               ))}
