+++ apps/webservice/src/app/[workspaceSlug]/(app)/(deploy)/(raw)/systems/[systemSlug]/(raw)/deployments/[deploymentSlug]/(sidebar)/_components/release-cell/policy-evaluation.ts
@@ -1,0 +1,9 @@
+export type PolicyEvaluationResult = {
+  policies: { id: string; name: string }[];
+  rules: {
+    anyApprovals: Record<string, string[]>;
+    roleApprovals: Record<string, string[]>;
+    userApprovals: Record<string, string[]>;
+    versionSelector: Record<string, boolean>;
+  };
+};
