+++ apps/webservice/src/app/[workspaceSlug]/(app)/(deploy)/(raw)/systems/[systemSlug]/(sidebar)/deployments/TableDeployments.tsx
@@ -230,7 +230,7 @@ const DeploymentTable: React.FC<{
                     <LazyDeploymentEnvironmentCell
                       environment={env}
                       deployment={r}
-                      systemSlug={systemSlug}
+                      system={{ slug: systemSlug }}
                     />
                   </td>
                 );
